# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PostBanner",
        "description": "API documentation for BannerApp",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/PostBanner/v1",
            "description": "V1 API"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login",
                "description": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "is_social",
                                    "login_with",
                                    "social_id",
                                    "device_token",
                                    "device_type",
                                    "uu_id"
                                ],
                                "properties": {
                                    "social_id": {
                                        "type": "string",
                                        "example": "google_id/apple_id"
                                    },
                                    "login_with": {
                                        "type": "string",
                                        "example": "Email/Google/Apple"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "cavin@mailinator.com"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Android"
                                    },
                                    "device_model": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Samsung SM-M146B - 13"
                                    },
                                    "device_os": {
                                        "type": "integer",
                                        "example": "os"
                                    },
                                    "app_version": {
                                        "type": "integer",
                                        "example": "version"
                                    },
                                    "api_version": {
                                        "type": "integer",
                                        "example": "api version"
                                    },
                                    "uu_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123"
                                    },
                                    "apple_auth_code": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123avcaca123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/logout/{uu_id}": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "logout",
                "description": "logout",
                "operationId": "logout",
                "parameters": [
                    {
                        "name": "uu_id",
                        "in": "path",
                        "description": "uu_id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/adddevicetoken": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "add device token",
                "description": "",
                "operationId": "adddevicetoken",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "uu_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "example": "Android"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "example": "token123"
                                    },
                                    "device_model": {
                                        "type": "string",
                                        "example": "SAMSUNG F12"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "name"
                                    },
                                    "device_os": {
                                        "type": "integer",
                                        "example": "os"
                                    },
                                    "app_version": {
                                        "type": "integer",
                                        "example": "version"
                                    },
                                    "api_version": {
                                        "type": "integer",
                                        "example": "api version"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Device token added successfully!"
                    },
                    "401": {
                        "description": "Something went wrong, Please try again."
                    }
                }
            }
        },
        "/verifyotp": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "verify OTP",
                "description": "verify OTP",
                "operationId": "verifyotp",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email_otp_number",
                                    "email",
                                    "mobile"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "emailId",
                                        "example": "stiv@mailinator.com"
                                    },
                                    "is_social": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "otp_number": {
                                        "type": "integer",
                                        "example": "5648"
                                    },
                                    "device_token": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123"
                                    },
                                    "device_type": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Android"
                                    },
                                    "device_model": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Samsung SM-M146B - 13"
                                    },
                                    "device_os": {
                                        "type": "integer",
                                        "example": "os"
                                    },
                                    "app_version": {
                                        "type": "integer",
                                        "example": "version"
                                    },
                                    "api_version": {
                                        "type": "integer",
                                        "example": "api version"
                                    },
                                    "uu_id": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sorry, wrong email address. Please try again"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/userdetails": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Sign in",
                "description": "get logged user details",
                "operationId": "userdetails",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "Something went wrong, Please try again."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/deleteuser": {
            "delete": {
                "tags": [
                    "Setting"
                ],
                "summary": "delete user",
                "description": "delete user",
                "operationId": "deleteuser",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "401": {
                        "description": "Something went wrong, Please try again."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}